break up robot2022 into subsystems and commands packages
