+++ src/main/kotlin/frc/team449/control/holonomic/SwerveDrive.kt
@@ -15,7 +15,7 @@
  import edu.wpi.first.wpilibj2.command.SubsystemBase
  import frc.team449.robot2022.constants.RobotConstants
  import frc.team449.robot2022.constants.SwerveConstants
- import frc.team449.robot2022.vision.VisionConstants
+ import frc.team449.robot2022.constants.VisionConstants
  import frc.team449.system.AHRS
  import frc.team449.system.VisionCamera
  import frc.team449.system.encoder.AbsoluteEncoder
