+++ src/main/kotlin/frc/team449/robot2022/auto/Example.kt
@@ -6,7 +6,7 @@ import edu.wpi.first.wpilibj2.command.Command
 import edu.wpi.first.wpilibj2.command.PrintCommand
 import frc.team449.control.auto.HolonomicRoutine
 import frc.team449.robot2022.Robot
-import frc.team449.robot2022.drive.HeadingAlign
+import frc.team449.robot2022.commands.HeadingAlign
 
 class Example(
   private val robot: Robot
